+++ src/ClockifyCli/Models/StartTimeEntry.cs
@@ -6,6 +6,6 @@ public record StartTimeEntry(
     [property: JsonProperty("start")] string Start,
     [property: JsonProperty("projectId")] string ProjectId,
     [property: JsonProperty("taskId")] string? TaskId,
-    [property: JsonProperty("description")]
-    string? Description
+    [property: JsonProperty("description")] string? Description,
+    [property: JsonProperty("type")] string Type
 );